GitWrap/Errors: Rename Git.{h,cpp} to GitWrap.*
