+++ src/Libs/GitWrap/BackgroundFetch.h
@@ -17,7 +17,7 @@
 #ifndef GIT_BACKGROUND_FETCH_H
 #define GIT_BACKGROUND_FETCH_H
 
-#include "Git.h"
+#include "GitWrap.h"
 #include "BackgroundTasks.h"
 #include "Repository.h"
 #include "Remote.h"
