+++ src/Libs/GitWrap/BackgroundTasks.h
@@ -21,7 +21,7 @@
 #include <QQueue>
 #include <QMutex>
 
-#include "Git.h"
+#include "GitWrap.h"
 
 namespace Git
 {
