+++ src/Libs/GitWrap/Config.h
@@ -14,7 +14,7 @@
  *
  */
 
-#include "Git.h"
+#include "GitWrap.h"
 
 namespace Git
 {
