+++ src/Libs/GitWrap/ObjectBlob.h
@@ -17,7 +17,7 @@
 #ifndef GIT_OBJECT_BLOB_H
 #define GIT_OBJECT_BLOB_H
 
-#include "Git.h"
+#include "GitWrap.h"
 #include "ObjectId.h"
 #include "Object.h"
